requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.120.1
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.3
pydantic_core==2.41.4
python-telegram-bot==20.6
requests==2.32.5
sniffio==1.3.1
starlette==0.49.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
